fix bogus setMargin in pyside2-uic
This bug has been reported five years ago.

By inspecting uiparser.py, in function createLayout, an extra 'margin' element is
added (~line 399) because the layout is child of a QWidget which is not the child
of the main window. The code in that function checks if there are individual
margin elements but fails in the case where there is a single 'margin' element,
that is, when all margins are equal. If at least one margin is different then the
problem doesn't appear.

Task-number: PYSIDE-127
Change-Id: I557a12a15048c5c713badc44309e2bb2144a35a9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>